requirements.txt
SQLAlchemy~=2.0.30
fastapi~=0.111.0
pydantic~=2.7.1
PyJWT~=2.8.0
python-dotenv~=1.0.1
uvicorn~=0.29.0
starlette~=0.37.2